add patch for multiple versions, add build script
